pyproject.toml
[project]
name = "adb-rag-1"
version = "0.1.0"
description = "제약 SOP/QMS 문서 기반 RAG 시스템 (Azure Databricks)"
requires-python = ">=3.13"
dependencies = [
    # Databricks
    "databricks-sdk>=0.40.0,<0.41",
    "databricks-vectorsearch>=0.8.0,<0.9",
    "mlflow>=2.19.0,<3",
    # PDF 처리
    "pymupdf>=1.25.0,<1.26",
    "langchain-text-splitters>=0.3.0,<0.4",
    # 설정 및 유틸
    "pydantic>=2.10.0,<3",
    "pyyaml>=6.0.0,<7",
    "python-dotenv>=1.0.0,<2",
    # HTTP 클라이언트 (임베딩/LLM 직접 호출)
    "httpx>=0.28.0,<0.29",
    # 프론트엔드 (로컬 확인용)
    "gradio>=5.0.0,<6",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0,<0.10",
    "pyright>=1.1.390,<2",
    "pre-commit>=4.0.0,<5",
    "pytest>=8.3.0,<9",
    "pytest-asyncio>=0.25.0,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ─────────────────────────────────────────
# ruff 설정
# ─────────────────────────────────────────
[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
]
ignore = [
    "E501",  # line-length는 ruff format에 위임
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ─────────────────────────────────────────
# pytest 설정
# ─────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
